pyproject.toml
[project]
name = "codecraft"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "langchain>=0.3.18",
    "langchain-community>=0.3.17",
    "langchain-huggingface>=0.1.2",
    "langchain-ollama>=0.2.3",
    "langchain-weaviate>=0.0.4",
    "opencv-python>=4.11.0.86",
    "pdf2image>=1.17.0",
    "pillow>=11.1.0",
    "pymupdf>=1.25.3",
    "pypdf>=5.3.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "reportlab>=4.3.1",
    "sentence-transformers>=3.4.1",
    "tiktoken>=0.9.0",
    "uvicorn>=0.34.0",
    "weaviate-client>=4.11.0",
]
